https://trello.com/c/wacVyxYG Redid neural network, changed remaining size_t to int in Matrix, bugfixes

+++ src/Matrix.cpp
@@ -36,11 +36,11 @@ Matrix::~Matrix(){
 
 }
 
-size_t Matrix::getWidth() {
+int Matrix::getWidth() {
     return width;
 }
 
-size_t Matrix::getHeight() {
+int Matrix::getHeight() {
     return height;
 }
 
@@ -61,9 +61,9 @@ float Matrix::getMin() {
 }
 
 std::ostream& operator<<(std::ostream &strm, const Matrix &m) {
-    for(size_t y = 0; y < m.height; y++) {
+    for (int y = 0; y < m.height; y++) {
         strm << "[";
-        for(size_t x = 0; x < m.width - 1; x++) {
+        for (int x = 0; x < m.width - 1; x++) {
             strm << m.data[y * m.width + x] << ", ";
         }
         if(m.width != 0) strm << m.data[(y + 1) * m.width - 1];
@@ -76,8 +76,8 @@ std::ostream& operator<<(std::ostream &strm, const Matrix &m) {
 
 Matrix& Matrix::operator+=(const Matrix &rhs) {
     if(width != rhs.width || height != rhs.height) throw std::invalid_argument("Matrix dimensions (" + std::to_string(height) + " x " + std::to_string(width) + ") and (" + std::to_string(rhs.height) + " x " + std::to_string(rhs.width) + ") do not match.");
-    size_t s = width * height;
-    for(size_t i = 0; i < s; i++) {
+    int s = width * height;
+    for (int i = 0; i < s; i++) {
         data[i] += rhs.data[i];
     }
     return *this;
@@ -90,8 +90,8 @@ Matrix operator+(Matrix lhs, const Matrix& rhs) {
 
 Matrix& Matrix::operator-=(const Matrix &rhs) {
     if(width != rhs.width || height != rhs.height) throw std::invalid_argument("Matrix dimensions (" + std::to_string(height) + " x " + std::to_string(width) + ") and (" + std::to_string(rhs.height) + " x " + std::to_string(rhs.width) + ") do not match.");
-    size_t s = width * height;
-    for(size_t i = 0; i < s; i++) {
+    int s = width * height;
+    for (int i = 0; i < s; i++) {
         data[i] -= rhs.data[i];
     }
     return *this;
@@ -104,8 +104,8 @@ Matrix operator-(Matrix lhs, const Matrix& rhs) {
 
 Matrix& Matrix::operator*=(const Matrix &rhs) {
     if(width != rhs.width || height != rhs.height) throw std::invalid_argument("Matrix dimensions (" + std::to_string(height) + " x " + std::to_string(width) + ") and (" + std::to_string(rhs.height) + " x " + std::to_string(rhs.width) + ") do not match.");
-    size_t s = width * height;
-    for(size_t i = 0; i < s; i++) {
+    int s = width * height;
+    for (int i = 0; i < s; i++) {
         data[i] *= rhs.data[i];
     }
     return *this;
@@ -118,8 +118,8 @@ Matrix operator*(Matrix lhs, const Matrix& rhs) {
 
 Matrix& Matrix::operator/=(const Matrix &rhs) {
     if(width != rhs.width || height != rhs.height) throw std::invalid_argument("Matrix dimensions (" + std::to_string(height) + " x " + std::to_string(width) + ") and (" + std::to_string(rhs.height) + " x " + std::to_string(rhs.width) + ") do not match.");
-    size_t s = width * height;
-    for(size_t i = 0; i < s; i++) {
+    int s = width * height;
+    for (int i = 0; i < s; i++) {
         data[i] /= rhs.data[i];
     }
     return *this;
@@ -133,7 +133,7 @@ Matrix operator/(Matrix lhs, const Matrix& rhs) {
 //########################### Matrix/float operators ###########################
 
 Matrix& Matrix::operator+=(float rhs) {
-    size_t s = width * height;
+    int s = width * height;
     for(int i = 0; i < s; i++) {
         data[i] += rhs;
     }
@@ -146,7 +146,7 @@ Matrix operator+(Matrix lhs, float rhs) {
 }
 
 Matrix& Matrix::operator-=(float rhs) {
-    size_t s = width * height;
+    int s = width * height;
     for(int i = 0; i < s; i++) {
         data[i] -= rhs;
     }
@@ -159,7 +159,7 @@ Matrix operator-(Matrix lhs, float rhs) {
 }
 
 Matrix& Matrix::operator*=(float rhs) {
-    size_t s = width * height;
+    int s = width * height;
     for(int i = 0; i < s; i++) {
         data[i] *= rhs;
     }
@@ -172,7 +172,7 @@ Matrix operator*(Matrix lhs, float rhs) {
 }
 
 Matrix& Matrix::operator/=(float rhs) {
-    size_t s = width * height;
+    int s = width * height;
     for(int i = 0; i < s; i++) {
         data[i] /= rhs;
     }
@@ -185,12 +185,12 @@ Matrix operator/(Matrix lhs, float rhs) {
 }
 //########################### Matrix Subscript ###########################
 
-float& Matrix::operator()(size_t y, size_t x) {
+float &Matrix::operator()(int y, int x) {
     if(x >= width || y >= height) throw std::invalid_argument("The indices (" + std::to_string(y) + ", " + std::to_string(x) + ") are invalid for a Matrix of size ("  + std::to_string(height) + " x " + std::to_string(width) + ").");
     return data[y * width + x];
 }
 
-float Matrix::operator()(size_t y, size_t x) const {
+float Matrix::operator()(int y, int x) const {
     if(x >= width || y >= height) throw std::invalid_argument("The indices (" + std::to_string(y) + ", " + std::to_string(x) + ") are invalid for a Matrix of size ("  + std::to_string(height) + " x " + std::to_string(width) + ").");
     return data[y * width + x];
 }
@@ -199,8 +199,8 @@ float Matrix::operator()(size_t y, size_t x) const {
 
 bool operator==(const Matrix &lhs, const Matrix &rhs) {
     if(lhs.width != rhs.width || lhs.height != rhs.height) return false;
-    size_t s = lhs.width * lhs.height;
-    for(size_t i = 0; i < s; i++) {
+    int s = lhs.width * lhs.height;
+    for (int i = 0; i < s; i++) {
         if(!(fabs(lhs.data[i] - rhs.data[i]) < lhs.data[i] * 0.001)) return false;
     }
     return true;
@@ -209,9 +209,9 @@ bool operator==(const Matrix &lhs, const Matrix &rhs) {
 Matrix Matrix::dotProduct(const Matrix &other) {
     if(width != other.height) throw std::invalid_argument("Width of the first Matrix with dimensions (" + std::to_string(height) + " x " + std::to_string(width) + ") must be equal to the height of the second Matrix with dimensions (" + std::to_string(other.height) + " x " + std::to_string(other.width) + ").");
     Matrix m(height, other.width);
-    for(size_t x2 = 0; x2 < other.width; x2++) {
-        for(size_t y1 = 0; y1 < height; y1++) {
-            for(size_t x1 = 0; x1 < width; x1++) {
+    for (int x2 = 0; x2 < other.width; x2++) {
+        for (int y1 = 0; y1 < height; y1++) {
+            for (int x1 = 0; x1 < width; x1++) {
                 m.data[y1 * m.width + x2] += data[y1 * width + x1] * other.data[x1 * other.width + x2];
             }
         }
@@ -220,7 +220,7 @@ Matrix Matrix::dotProduct(const Matrix &other) {
 }
 
 Matrix& Matrix::apply(MatrixFunction fun) {
-    size_t s = width * height;
+    int s = width * height;
     for(int i = 0; i < s; i++) {
         data[i] = fun(data[i]);
     }
@@ -238,8 +238,8 @@ float Matrix::toFloat() {
 
 Matrix& Matrix::transpose() {
     if(width == height) {
-        for(size_t x = 0; x < width - 1; x++) {
-            for(size_t y = x + 1; y < width; y++) { // y = x + 1 -> always takes all values below the "middle line"
+        for (int x = 0; x < width - 1; x++) {
+            for (int y = x + 1; y < width; y++) { // y = x + 1 -> always takes all values below the "middle line"
                 //swap data[x, y] and data[y, x]
                 float temp = data[y * width + x];
                 data[y * width + x] = data[x * width + y];
@@ -248,13 +248,13 @@ Matrix& Matrix::transpose() {
         }
     } else { //inefficient af, because I copy it instead of calculating permutation circles
         std::vector<float> dataNew(width * height);
-        for(size_t y = 0; y < height; y++) {
-	        for(size_t x = 0; x < width; x++) {
+        for (int y = 0; y < height; y++) {
+            for (int x = 0; x < width; x++) {
 	        	dataNew[x * height + y] = data[y * width + x];
             }
         }
         data = dataNew;
-        size_t w = width;
+        int w = width;
         width = height;
         height = w;
     }
